% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tweets.R
\name{download_tweets}
\alias{download_tweets}
\title{Get tweets for given statuses (tweet IDs).}
\usage{
download_tweets(consumer_key, consumer_secret, access_token, access_secret,
  status_ids)
}
\arguments{
\item{consumer_key}{Consumer Key (API Key) from https://apps.twitter.com/}

\item{consumer_secret}{Consumer Secret (API Secret) from https://apps.twitter.com/}

\item{access_token}{Access Token from the https://apps.twitter.com/}

\item{access_secret}{Access Token Secret from https://apps.twitter.com/}

\item{status_ids}{data frame of tweet IDs}
}
\value{
A tibble of tweets data.
}
\description{
Get tweets for given statuses (tweet IDs).
}
